+++ TypeContractor/Output/DestinationType.cs
@@ -1,8 +1,8 @@
 namespace TypeContractor.Output;
 
-public record DestinationType(string TypeName, string ImportType, bool IsBuiltin, bool IsArray, Type? InnerType)
+public record DestinationType(string TypeName, string ImportType, bool IsBuiltin, bool IsArray, bool IsReadonly, Type? InnerType)
 {
-    public DestinationType(string typeName, bool isBuiltin, bool isArray, Type? innerType, string? importType = null) : this(typeName, importType ?? typeName, isBuiltin, isArray, innerType)
+    public DestinationType(string typeName, bool isBuiltin, bool isArray, bool isReadonly, Type? innerType, string? importType = null) : this(typeName, importType ?? typeName, isBuiltin, isArray, isReadonly, innerType)
     {
     }